OTel annotations
